Include fail tests and remove LUT implementation
Commit of tests to make the algorithms fail along with some ideas,
how to fix the problems. There was also a missuse of 'constexpr'
in the code. Seems, that using this keyword to calculate values
while compile time needs more effort and is less straightforword
than expected. Possbile implementation would be the use of an exernal
file, which is set up meanwhile the first run, but is this convenient?