Add new cycle measurement and run measurements
Time and cycles can be measured now and the tests for all the algorithms
has been carried out.

More Importent open questions:
- Can there be a different amount of cycles for the same set of instructions?
- How can I measure the correctnes of the cycle measure function?
	* At least as a reference we should implement the CPUID + RDTSC
	  methode.

Please note, that the makefile now contains a new flag, of the CPU supports
RDTSCP. This flag is now set manually. The Core i5 does support this
instruction.