Add new fib.algorithm nr7 & new benchmark methodes
Lots of changes has been done:
	- A new fibonacci algorithm has been implemented, using
	  exponentiation by squaring in an iterative manner
	- the cycle measuring function is now more stabil by using
	  the cpuid instruction to prevent the moving instructions
	  before or after the cycle reading into the measuring code
	- the benchmark methode has been design to be more abstract;
	  therefore it can be used to perform different benchmark
	  measurings without having a lot code overhead
	- exporting the measurements into a file has been moved into
	  an own function
	- the matrix class has been extended to provide new functions

Nevertheless the measuring could be much more easy. Most of the problems
arose from the writing methods, which are used to write into the files.
Most likely it would be worth to switch from printf to the streambased
'>>' operator. It would be easier to handly the different measuring units
and also compiling the header is most likely easier.